backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.2
djoser==2.1.0
drf-extra-fields==3.7.0
gunicorn==20.1.0
# webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
PyYAML==6.0
python-dotenv==1.0.0

